Refactor PaletteGroup to PaletteGroup
