+++ foundry/core/player_animations/util.py
@@ -1,6 +1,7 @@
 from foundry.core.graphics_set.GraphicsSet import GraphicalPage, GraphicsSet
 from foundry.core.palette import COLORS_PER_PALETTE
-from foundry.core.palette.Palette import Palette, PaletteGroup, PaletteGroupProtocol
+from foundry.core.palette.Palette import Palette
+from foundry.core.palette.PaletteGroup import PaletteGroup, PaletteGroupProtocol
 from foundry.core.player_animations import (
     PLAYER_FRAME_PAGE_OFFSET,
     PLAYER_FRAME_START,
