+++ foundry/game/gfx/objects/EnemyItem.py
@@ -6,7 +6,7 @@ from foundry.core.graphics_set.GraphicsSet import (
     GraphicsSet,
     GraphicsSetProtocol,
 )
-from foundry.core.palette.Palette import PaletteGroup
+from foundry.core.palette.PaletteGroup import PaletteGroup
 from foundry.core.Position import Position, PositionProtocol
 from foundry.game.EnemyDefinitions import (
     EnemyDefinition,
