+++ foundry/game/gfx/objects/EnemyItemFactory.py
@@ -3,7 +3,7 @@ from PySide6.QtGui import QImage
 
 from foundry import data_dir
 from foundry.core.palette import PALETTE_GROUPS_PER_OBJECT_SET
-from foundry.core.palette.Palette import PaletteGroup
+from foundry.core.palette.PaletteGroup import PaletteGroup
 from foundry.game.gfx.drawable.Block import Block
 from foundry.game.gfx.objects.EnemyItem import EnemyObject
