+++ foundry/gui/PaletteGroupModel.py
@@ -7,7 +7,7 @@ from foundry.core.palette import (
     PALETTE_GROUPS_PER_OBJECT_SET,
     PALETTES_PER_PALETTES_GROUP,
 )
-from foundry.core.palette.Palette import PaletteGroup, PaletteGroupProtocol
+from foundry.core.palette.PaletteGroup import PaletteGroup, PaletteGroupProtocol
 from foundry.core.palette.util import get_internal_palette_offset
 from foundry.game.File import ROM
