+++ foundry/gui/PlayerViewer.py
@@ -15,12 +15,8 @@ from PySide6.QtWidgets import (
 )
 
 from foundry import icon
-from foundry.core.palette.Palette import (
-    Palette,
-    PaletteGroup,
-    PaletteGroupProtocol,
-    PaletteProtocol,
-)
+from foundry.core.palette.Palette import Palette, PaletteProtocol
+from foundry.core.palette.PaletteGroup import PaletteGroup, PaletteGroupProtocol
 from foundry.core.player_animations import ANIMATION_WIDTH
 from foundry.core.player_animations.PlayerAnimation import PlayerAnimation
 from foundry.core.player_animations.util import (
